+++ documentdb_tests/compatibility/tests/system/diagnostic/commands/validate/test_validate_core_behavior.py
@@ -1,23 +1,26 @@
 """Tests for validate command core behavior.
 
-Validates basic functionality, counts, consistency across calls, and comment parameter.
+Validates basic functionality, counts, consistency across calls, comment parameter,
+behavior on different collection types (capped, timeseries, clustered), and valid
+option combinations.
 """
 
 from __future__ import annotations
+
+from datetime import datetime, timezone
 
 import pytest
 
 from documentdb_tests.compatibility.tests.system.diagnostic.utils.diagnostic_test_case import (
     DiagnosticTestCase,
 )
-from documentdb_tests.framework.assertions import assertFailureCode, assertProperties
-from documentdb_tests.framework.error_codes import NAMESPACE_NOT_FOUND_ERROR
+from documentdb_tests.framework.assertions import assertProperties, assertSuccessPartial
 from documentdb_tests.framework.executor import execute_command
 from documentdb_tests.framework.parametrize import pytest_params
 from documentdb_tests.framework.property_checks import Eq
 
-# Property [Core Behavior]: validate returns expected results for populated,
-# empty, and non-existent collections and supports common parameters.
+# Property [Core Behavior]: validate returns expected results for populated
+# and empty collections and supports common parameters.
 CORE_BEHAVIOR_TESTS: list[DiagnosticTestCase] = [
     DiagnosticTestCase(
         "populated_collection",
@@ -64,16 +67,6 @@ CORE_BEHAVIOR_TESTS: list[DiagnosticTestCase] = [
     ),
 ]
 
-# Property [Non-Existent Collection]: validate returns NamespaceNotFound for
-# a collection that does not exist.
-NON_EXISTENT_TESTS: list[DiagnosticTestCase] = [
-    DiagnosticTestCase(
-        "non_existent_collection",
-        error_code=NAMESPACE_NOT_FOUND_ERROR,
-        msg="validate should return NamespaceNotFound for a non-existent collection",
-    ),
-]
-
 
 @pytest.mark.parametrize("test", pytest_params(CORE_BEHAVIOR_TESTS))
 def test_validate_core_behavior(collection, test):
@@ -85,13 +78,6 @@ def test_validate_core_behavior(collection, test):
         cmd.update(test.command)
     result = execute_command(collection, cmd)
     assertProperties(result, test.checks, msg=test.msg, raw_res=True)
-
-
-@pytest.mark.parametrize("test", pytest_params(NON_EXISTENT_TESTS))
-def test_validate_non_existent(collection, test):
-    """Test validate on non-existent collections returns expected error."""
-    result = execute_command(collection, {"validate": f"{collection.name}_nonexistent_xyz"})
-    assertFailureCode(result, test.error_code, msg=test.msg)
 
 
 def test_validate_consistent_across_calls(collection):
@@ -123,3 +109,113 @@ def test_validate_reflects_modifications(collection):
         raw_res=True,
         msg="validate should reflect updated nrecords after additional inserts",
     )
+
+
+def test_validate_capped_collection(database_client, collection):
+    """Test validate on a capped collection succeeds."""
+    coll_name = f"{collection.name}_capped"
+    database_client.create_collection(coll_name, capped=True, size=1_048_576)
+    coll = database_client[coll_name]
+    coll.insert_one({"_id": 1, "x": 1})
+    result = execute_command(coll, {"validate": coll.name})
+    assertSuccessPartial(
+        result,
+        {"ok": 1.0, "valid": True},
+        msg="validate should succeed on a capped collection",
+    )
+
+
+def test_validate_timeseries_collection(database_client, collection):
+    """Test validate on a time series collection succeeds."""
+    coll_name = f"{collection.name}_timeseries"
+    database_client.create_collection(
+        coll_name,
+        timeseries={"timeField": "ts", "metaField": "meta"},
+    )
+    coll = database_client[coll_name]
+    coll.insert_one({"ts": datetime(2024, 1, 1, tzinfo=timezone.utc), "meta": "a", "v": 1})
+    result = execute_command(coll, {"validate": coll.name})
+    assertSuccessPartial(
+        result,
+        {"ok": 1.0},
+        msg="validate should succeed on a timeseries collection",
+    )
+
+
+def test_validate_clustered_collection(database_client, collection):
+    """Test validate on a clustered collection succeeds."""
+    coll_name = f"{collection.name}_clustered"
+    database_client.command(
+        "create",
+        coll_name,
+        clusteredIndex={"key": {"_id": 1}, "unique": True},
+    )
+    coll = database_client[coll_name]
+    coll.insert_one({"_id": 1, "x": 1})
+    result = execute_command(coll, {"validate": coll.name})
+    assertSuccessPartial(
+        result,
+        {"ok": 1.0},
+        msg="validate should succeed on a clustered collection",
+    )
+
+
+# Property [Valid Combinations]: validate succeeds with valid option combinations.
+VALID_COMBINATION_TESTS: list[DiagnosticTestCase] = [
+    DiagnosticTestCase(
+        "all_defaults_explicit",
+        command={"full": False, "repair": False, "metadata": False, "checkBSONConformance": False},
+        checks={"ok": Eq(1.0)},
+        msg="validate should succeed with all options set to false explicitly",
+    ),
+    DiagnosticTestCase(
+        "full_true",
+        command={"full": True},
+        checks={"ok": Eq(1.0)},
+        msg="validate with full: true should succeed",
+    ),
+    DiagnosticTestCase(
+        "checkBSONConformance_true",
+        command={"checkBSONConformance": True},
+        checks={"ok": Eq(1.0)},
+        msg="validate with checkBSONConformance: true should succeed",
+    ),
+    DiagnosticTestCase(
+        "full_and_checkBSONConformance",
+        command={"full": True, "checkBSONConformance": True},
+        checks={"ok": Eq(1.0)},
+        msg="validate with full: true and checkBSONConformance: true should succeed",
+    ),
+    DiagnosticTestCase(
+        "metadata_true",
+        command={"metadata": True},
+        checks={"ok": Eq(1.0)},
+        msg="validate with metadata: true should succeed",
+    ),
+    DiagnosticTestCase(
+        "fixMultikey_true_alone",
+        command={"fixMultikey": True},
+        checks={"ok": Eq(1.0)},
+        msg="validate with fixMultikey: true alone should succeed",
+    ),
+    DiagnosticTestCase(
+        "repair_true_alone",
+        command={"repair": True},
+        checks={"ok": Eq(1.0)},
+        msg="validate with repair: true alone should succeed",
+    ),
+    DiagnosticTestCase(
+        "repair_true_with_fixMultikey",
+        command={"repair": True, "fixMultikey": True},
+        checks={"ok": Eq(1.0)},
+        msg="validate with repair: true and fixMultikey: true should succeed",
+    ),
+]
+
+
+@pytest.mark.parametrize("test", pytest_params(VALID_COMBINATION_TESTS))
+def test_validate_valid_option_combinations(collection, test):
+    """Test that validate succeeds with valid option combinations."""
+    collection.insert_one({"_id": 1})
+    result = execute_command(collection, {"validate": collection.name, **test.command})
+    assertProperties(result, test.checks, msg=test.msg, raw_res=True)
